Begin refactoring editor views into a module
